+++ src/main/java/com/prgm/aroundthetown/room/entity/Room.java
@@ -1,4 +1,4 @@
-package com.prgm.aroundthetown.product.room.entity;
+package com.prgm.aroundthetown.room.entity;
 
 import com.prgm.aroundthetown.common.BaseEntity;
 import com.prgm.aroundthetown.product.accommodation.entity.Accommodation;
