+++ src/main/java/com/prgm/aroundthetown/room/repository/RoomImageRepository.java
@@ -1,6 +1,6 @@
-package com.prgm.aroundthetown.product.room.repository;
+package com.prgm.aroundthetown.room.repository;
 
-import com.prgm.aroundthetown.product.room.entity.RoomImage;
+import com.prgm.aroundthetown.room.entity.RoomImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomImageRepository extends JpaRepository<RoomImage, Long> {
