+++ src/main/java/com/prgm/aroundthetown/room/repository/RoomReservationRepository.java
@@ -1,6 +1,6 @@
-package com.prgm.aroundthetown.product.room.repository;
+package com.prgm.aroundthetown.room.repository;
 
-import com.prgm.aroundthetown.product.room.entity.RoomReservation;
+import com.prgm.aroundthetown.room.entity.RoomReservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomReservationRepository extends JpaRepository<RoomReservation, Long> {
