+++ src/main/java/com/prgm/aroundthetown/ticket/entity/Ticket.java
@@ -1,4 +1,4 @@
-package com.prgm.aroundthetown.product.ticket.entity;
+package com.prgm.aroundthetown.ticket.entity;
 
 import com.prgm.aroundthetown.common.BaseEntity;
 import com.prgm.aroundthetown.product.leisure.entity.Leisure;
