+++ src/main/java/com/prgm/aroundthetown/ticket/repository/TicketImageRepository.java
@@ -1,6 +1,6 @@
-package com.prgm.aroundthetown.product.ticket.repository;
+package com.prgm.aroundthetown.ticket.repository;
 
-import com.prgm.aroundthetown.product.ticket.entity.TicketImage;
+import com.prgm.aroundthetown.ticket.entity.TicketImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TicketImageRepository extends JpaRepository<TicketImage, Long> {
