+++ src/main/java/com/prgm/aroundthetown/ticket/repository/TicketRepository.java
@@ -1,6 +1,6 @@
-package com.prgm.aroundthetown.product.ticket.repository;
+package com.prgm.aroundthetown.ticket.repository;
 
-import com.prgm.aroundthetown.product.ticket.entity.Ticket;
+import com.prgm.aroundthetown.ticket.entity.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TicketRepository extends JpaRepository<Ticket, Long> {
