+++ src/test/java/com/prgm/aroundthetown/ticket/repository/TicketImageRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.prgm.aroundthetown.product.ticket.repository;
+package com.prgm.aroundthetown.ticket.repository;
 
 import com.prgm.aroundthetown.host.entity.Host;
 import com.prgm.aroundthetown.host.repository.HostRepository;
@@ -7,7 +7,8 @@ import com.prgm.aroundthetown.product.Region;
 import com.prgm.aroundthetown.product.leisure.entity.Leisure;
 import com.prgm.aroundthetown.product.leisure.entity.LeisureCategory;
 import com.prgm.aroundthetown.product.leisure.repository.LeisureRepository;
-import com.prgm.aroundthetown.product.ticket.entity.Ticket;
+import com.prgm.aroundthetown.ticket.entity.Ticket;
+import com.prgm.aroundthetown.ticket.entity.TicketImage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,8 +22,10 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
 @SpringBootTest
-class TicketRepositoryTest {
+class TicketImageRepositoryTest {
 
+    @Autowired
+    private TicketImageRepository ticketImageRepository;
     @Autowired
     private TicketRepository ticketRepository;
     @Autowired
@@ -30,7 +33,7 @@ class TicketRepositoryTest {
     @Autowired
     private HostRepository hostRepository;
 
-    private Leisure savedLeisure;
+    private Ticket savedTicket;
 
     @BeforeEach
     void setUp() {
@@ -56,23 +59,29 @@ class TicketRepositoryTest {
                 .expirationDate(LocalDateTime.now())
                 .category(LeisureCategory.AMUSEMENTPARK)
                 .build();
-        savedLeisure = leisureRepository.save(leisure);
+        final Leisure savedLeisure = leisureRepository.save(leisure);
+
+        final Ticket ticket = Ticket.builder()
+                .ticketName("2인용")
+                .price(25000)
+                .leisure(savedLeisure)
+                .build();
+        savedTicket = ticketRepository.save(ticket);
     }
 
     @Test
     @DisplayName("ticket을 save 할 수 있다.")
     @Transactional
     void saveTicketTest() {
-        final Ticket ticket = Ticket.builder()
-                .ticketName("2인용")
-                .price(25000)
-                .leisure(savedLeisure)
+        final TicketImage image = TicketImage.builder()
+                .IMAGE_PATH("path")
+                .ticket(savedTicket)
                 .build();
-        ticketRepository.save(ticket);
+        ticketImageRepository.save(image);
 
-        assertThat(ticketRepository.findAll().size(), is(1));
+        assertThat(ticketImageRepository.findAll().size(), is(1));
 
         // 연관관계 mapping
-        assertThat(leisureRepository.findAll().get(0).getTickets().size(), is(1));
+        assertThat(ticketRepository.findAll().get(0).getTicketImages().size(), is(1));
     }
 }
