+++ src/test/java/com/prgm/aroundthetown/ticket/repository/TicketRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.prgm.aroundthetown.product.ticket.repository;
+package com.prgm.aroundthetown.ticket.repository;
 
 import com.prgm.aroundthetown.host.entity.Host;
 import com.prgm.aroundthetown.host.repository.HostRepository;
@@ -7,8 +7,7 @@ import com.prgm.aroundthetown.product.Region;
 import com.prgm.aroundthetown.product.leisure.entity.Leisure;
 import com.prgm.aroundthetown.product.leisure.entity.LeisureCategory;
 import com.prgm.aroundthetown.product.leisure.repository.LeisureRepository;
-import com.prgm.aroundthetown.product.ticket.entity.Ticket;
-import com.prgm.aroundthetown.product.ticket.entity.TicketImage;
+import com.prgm.aroundthetown.ticket.entity.Ticket;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -22,10 +21,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
 @SpringBootTest
-class TicketImageRepositoryTest {
+class TicketRepositoryTest {
 
-    @Autowired
-    private TicketImageRepository ticketImageRepository;
     @Autowired
     private TicketRepository ticketRepository;
     @Autowired
@@ -33,7 +30,7 @@ class TicketImageRepositoryTest {
     @Autowired
     private HostRepository hostRepository;
 
-    private Ticket savedTicket;
+    private Leisure savedLeisure;
 
     @BeforeEach
     void setUp() {
@@ -59,29 +56,23 @@ class TicketImageRepositoryTest {
                 .expirationDate(LocalDateTime.now())
                 .category(LeisureCategory.AMUSEMENTPARK)
                 .build();
-        final Leisure savedLeisure = leisureRepository.save(leisure);
-
-        final Ticket ticket = Ticket.builder()
-                .ticketName("2인용")
-                .price(25000)
-                .leisure(savedLeisure)
-                .build();
-        savedTicket = ticketRepository.save(ticket);
+        savedLeisure = leisureRepository.save(leisure);
     }
 
     @Test
     @DisplayName("ticket을 save 할 수 있다.")
     @Transactional
     void saveTicketTest() {
-        final TicketImage image = TicketImage.builder()
-                .IMAGE_PATH("path")
-                .ticket(savedTicket)
+        final Ticket ticket = Ticket.builder()
+                .ticketName("2인용")
+                .price(25000)
+                .leisure(savedLeisure)
                 .build();
-        ticketImageRepository.save(image);
+        ticketRepository.save(ticket);
 
-        assertThat(ticketImageRepository.findAll().size(), is(1));
+        assertThat(ticketRepository.findAll().size(), is(1));
 
         // 연관관계 mapping
-        assertThat(ticketRepository.findAll().get(0).getTicketImages().size(), is(1));
+        assertThat(leisureRepository.findAll().get(0).getTickets().size(), is(1));
     }
 }
